command fully automated

+++ core/management/commands/get_artwork.py
@@ -22,26 +22,30 @@ class Command(BaseCommand):
             pokemon_list = response.json()
             total_count = pokemon_list["count"]
 
-            for item in pokemon_list["results"]:
-                pokemon_response = requests.get(item["url"])
+            # for item in pokemon_list["results"]:
+            for i in range(total_count):
+                list_item = pokemon_list["results"][i]
+                pokemon_response = requests.get(list_item["url"])
                 if pokemon_response.status_code == 200:
                     pokemon = pokemon_response.json()
-                    print(pokemon["name"])
+                    pokemon_id = str(pokemon["id"])
+                    print(pokemon_id + ": " + pokemon["name"])
 
                     """Download the artwork for each pokemon and store it in the created directory"""
-                    artwork_url = pokemon["sprites"]["other"]["official-artwork"]["front_default"]
-                    image = artwork_path + str(pokemon["id"]) + ".png"
-                    artwork_response = requests.get(artwork_url)
+                    artwork_url = f"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/{pokemon_id}.png"
+                    image = artwork_path + pokemon_id + ".png"
+                    image_path = Path(image)
 
-                    if artwork_response.status_code == 200:
-                        file = open(image, "wb")
-                        file.write(artwork_response.content)
-                        file.close
-                        print(image + " has successfully downloaded.")
-                    elif response.status_code == 404:
-                        print(f"{image} does not exist.")                
-                    else:
-                        print("Error. Could not retrieve" + image)
-                    
-
+                    if not image_path.exists():      
+                        artwork_response = requests.get(artwork_url)
+                        if artwork_response.status_code == 200:
+                            file = open(image, "wb")
+                            file.write(artwork_response.content)
+                            file.close
+                            print(image + " has successfully downloaded.")
+                        elif response.status_code == 404:
+                            print(f"{image} does not exist.")                
+                        else:
+                            print("Error. Could not retrieve" + image)
+                        
         print("Done.")
